requirements.txt
torch
torchvision
timm
gradio
opencv-python-headless
numpy
Pillow
grad-cam